Changed view of space blueprints

+++ src/net/sourceforge/fenixedu/applicationTier/Servico/space/DeleteMaterialSpaceOccupation.java
@@ -6,6 +6,8 @@ import net.sourceforge.fenixedu.domain.space.MaterialSpaceOccupation;
 public class DeleteMaterialSpaceOccupation<T extends MaterialSpaceOccupation> extends Service {
 
     public void run(T t) {
-        t.delete();
+	if(t != null) {
+	    t.delete();
+	}
     }
 }
